+++ build.gradle.kts
@@ -1,21 +1,17 @@
 plugins {
-    kotlin("jvm") version "2.0.0"
+    kotlin("jvm") apply false
 }
 
 group = "ru.otus.otuskotlin.finaccounting"
 version = "1.0-SNAPSHOT"
 
-repositories {
-    mavenCentral()
+allprojects {
+    repositories {
+        mavenCentral()
+    }
 }
 
-dependencies {
-    testImplementation(kotlin("test"))
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
-kotlin {
-    jvmToolchain(21)
+subprojects {
+    group = rootProject.group
+    version = rootProject.version
 }
