+++ m1l1-first/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    kotlin("jvm")
+}
+
+dependencies {
+    testImplementation(kotlin("test-junit"))
+}
+
+kotlin {
+    jvmToolchain(21)
+}
