+++ m1l1-first/src/main/kotlin/Main.kt
@@ -1,0 +1,10 @@
+package ru.otus.otuskotlin.finaccounting
+
+fun main() {
+    val name = "Kotlin"
+    println("Hello, $name!")
+
+    for (i in 1..5) {
+        println("i = $i")
+    }
+}
