+++ settings.gradle.kts
@@ -1,5 +1,12 @@
+pluginManagement {
+    val kotlinVersion: String by settings
+    plugins {
+        kotlin("jvm") version kotlinVersion apply false
+    }
+}
+
 plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.8.0"
 }
 rootProject.name = "ok-finaccounting-202408"
-
+include("m1l1-first")
